+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Release script
 ### Changed
+- Decoration follows the cursor line
 
 ## v.0.0.1